+++ plugins/web_adapter/lib/dio_web_adapter.dart
@@ -1,7 +1,7 @@
 library dio_web_adapter;
 
-export 'src/adapter.dart';
-export 'src/compute.dart';
+export 'src/adapter_impl.dart';
+export 'src/compute_impl.dart';
 export 'src/dio_impl.dart';
-export 'src/multipart_file.dart';
-export 'src/progress_stream.dart';
+export 'src/multipart_file_impl.dart';
+export 'src/progress_stream_impl.dart';
